+++ Wire-iOS/Sources/UserInterface/Settings/ChangingDetails/ChangeHandleViewController.swift
@@ -185,7 +185,7 @@ struct HandleChangeState {
 }
 
 
-@objcMembers final class ChangeHandleViewController: SettingsBaseTableViewController {
+final class ChangeHandleViewController: SettingsBaseTableViewController {
 
     public var footerFont: UIFont = .smallFont
     var state: HandleChangeState
